refactor(app-shell): code split navbar but not navbar loading placeholder
