+++ packages/application-shell/src/components/application-shell-authenticated/application-shell-authenticated.tsx
@@ -34,7 +34,8 @@ import FetchProject from '../fetch-project';
 import FetchUser from '../fetch-user';
 import GtmApplicationTracker from '../gtm-application-tracker';
 import GtmUserTracker from '../gtm-user-tracker';
-import NavBar, { LoadingNavBar } from '../navbar';
+import NavBar from '../navbar';
+import LoadingNavBar from '../navbar/loading-navbar';
 import ProjectContainer from '../project-container';
 import ProjectDataLocale from '../project-data-locale';
 import QuickAccess from '../quick-access';
