+++ packages/application-shell/src/components/navbar/index.ts
@@ -1,1 +1,5 @@
-export { default, LoadingNavBar } from './navbar';
+import { lazy } from 'react';
+
+const Navbar = lazy(() => import('./navbar' /* webpackChunkName: "navbar" */));
+
+export default Navbar;
